move pages out of explorer
